Disable livereload in parent frame if present
When using a tool like [Viewport Resizer][], livereload is first
injected in the main window, then when we transform this window into an
embedded iframe, we get livereload loaded inside the iframe and outside
the iframe. This means that we will reload both documents on HTML change
and we will get kicked outside our iframe when the main window reload.

This makes using [Viewport Resizer][] quite difficult. The solution
proposed here is to disable liveloader in the parent frame if
present. Only the inner iframe will keep livereloader.

[Viewport Resizer]: http://lab.maltewassermann.com/viewport-resizer/